RD-4466 Local: default attributes of dep-updates
When creating a dep-update, default the structure of the object
prior to storing it. Then, the caller doesn't need to pass all of this.

Also, return the created object.

Also also, when creating a deployment, also return the created env.
That one's mostly unrelated but makes for nicer caller code.