RD-5109 Remove unnecessary stub subgraphs
Adding an optimize method to the tasks graph, which will remove all
the empty tasks from the graph.

Currently, let's only run this optimization in the "lifecycle"
workflows. Users are free to run it as they see fit.

Also, in the future, we might come up with more "optimizations".
This impl. right now is pretty straightforward, but feel free to later
rewrite it as something more complex.