specialcase cancellation
since I now made the Errors class format its own tracebacks, it will
not take exceptions as input anymore; so let's not pass a Canceled
exception there, but handle it explicitly here.

Now we have a choice of whether the `if self._errors` comes first,
or if `if api.has_cancel_request()` comes first - ie. if there's
both an error, and a cancel request, which should be the result?
Here, I decide that it's the error.