RD-5089 Keep running the subgraphs as far as possible (#1104)
* RD-5089 Keep running the subgraphs as far as possible

* add a test

this is the scenario we want to achieve here!

* remove task before adding a retry

this way handles ._ready correctly - now important because we're now
continuing after an error.

Actually I'm not sure I understand this fully, but this way passes the
tests. Hmm.

* if there's ready tasks, we're not finished yet

* start counting deadline from the first error

...not the last one. I like this behaviour better. If any failure
resets the timer, it's really hard to estimate how long is the workflow
still going to take.

* _task_error: move formatting the traceback into TaskGraphErrors

just delegating this part, to the class. The class doesn't yet
implement it, but it will, in a later commit.

* specialcase cancellation

since I now made the Errors class format its own tracebacks, it will
not take exceptions as input anymore; so let's not pass a Canceled
exception there, but handle it explicitly here.

Now we have a choice of whether the `if self._errors` comes first,
or if `if api.has_cancel_request()` comes first - ie. if there's
both an error, and a cancel request, which should be the result?
Here, I decide that it's the error.

* when we run out of retries, remove the retried task

Wasn't a problem before, because setting status to failed, just stopped
the execution. Now, it continues (in other subgraphs), so we must make
sure that this subgraph doesn't get retried anymore

* TaskDependencyGraphErrors: format exceptions

This stores all the exception data, and later on, formats an
exception compiled from all of them. Lots of twiddling to get the
output just the way I like it. It looks nice. But essentially this
is just formatting user-readable strings.

* Also wrap exceptions for operations, not just workflows

This will also allow us to hide the framework-specific lines from
operation tracebacks, same as we do with workflow tracebacks

* flake

Co-authored-by: asdf <tehasdf@gmail.com>