`@workflow_test`: allow passing blueprint contents as string
I don't want to always write tests by putting a blueprint in a separate
file, I want to just write it in a string right there! It's going
to make writing tests a lot easier.

Deciding whether the passed-in string is a filename or a blueprint
string is perhaps a bit stupid, but hey, this works.

Some care is taken to make sure all temp files are deleted indeed.