RD-6226 Optimize all tasks graphs when storing them (#1194)
* heal: consider missing result same as empty result

if 'status' is missing entirely, consider that the same as if 'status'
was present but 'task' was empty

* optimize all stored graphs

Let's goooo.

* when is nop not a nop?

NOPs with success/failure handlers aren't _REALLY_ nops, so we must
treat them as actual tasks. They do run some code!

* ...but if they're nops, don't attach handlers to them

check drift & status: no need to attach handlers to tasks that do
nothing. That way, the tasks will indeed be considered NOPs later on.
(for the purposes of optimizations)

* fixup tests

in that test, we used to assert on an instance that didn't have
the interface defined, but we still asserted on if the task actually
ran (or, its on_success handler). Now it doesn't run in that case,
but we'd still like to assert on that, so let's make that node
actually have that interface