test_builtin_workflows: more resilient write operations
well, when running many tasks (eg. a node check drift, and a
relationship check drift), we can try and concurrently write into
the same runtime-props, and that causes a version conflict.

To avoid that, we can use ctx.instance.update

It is a bit unwieldy, I wonder if there can be a better way!