Rd-6332 check_drift: default run_by_dependency_order to true (#1195)
* Rd-6332 check_drift: default run_by_dependency_order to true

This is the other part of cloudify-cosmo/cloudify-manager#3949

* test both variants of check_drift

with both dependency, and without. Of course, with dependencies,
we stop on failed graphs, and without dependencies, we just run
everything.

* test_builtin_workflows: more resilient write operations

well, when running many tasks (eg. a node check drift, and a
relationship check drift), we can try and concurrently write into
the same runtime-props, and that causes a version conflict.

To avoid that, we can use ctx.instance.update

It is a bit unwieldy, I wonder if there can be a better way!

* flake