dsl_parser tests: get rid of the warnings
Python 3.11 now reports a warning that a `return` in test cases is
deprecated.

In test_groups, we have tests which run asserts and then return the
plan, so that the plan is then tested further in "modify" tests.

But if the original test can't return stuff...

Instead, move the behaviour into separate methods, and have both
tests call that.

This is a simple mechanical change and allows me to not have to read
and understand all these tests! Woo.