RD-6537 client.agents.list(): include all agents (#1198)
* RD-6537 client.agents.list(): include all agents

I don't really see a reason to have this default filter. The original
ticket didn't mention it.
Possibly back then we were creating many failed agents?
But now, it's more stable, so let's just list all of them.

* agents list: also no need to constraint include

We maybe don't always want to fetch credentials, but a more modern way
to achieve that would be to use `get_data` - so that's what I'm
doing in cloudify-cosmo/cloudify-manager#3958

The problem with hardcoding the default include in here is that it
won't have any newer fields. Specifically, this one is missing `state`.