Rewrite the async client too
Now the async client is more of a thing of its own. Similar to how
you do CloudifyClient(), you can now also use AsyncCloudifyClient(),
and that is going to be very similar to the usual client, except..
well, async.

Not all endpoints are available yet. Some of them I couldn't easily
translate :(
And so, community_contacts, deployment_groups, and log_bundles, are
just not available for now. Still pretty good.