RD-6652 Scale: don't run unnecessary operations (#1219)
* Scale: make sure to refresh instances before doing anything

Very refreshing.
The various `lifecycle` methods examine instance state when scheduling
tasks, so the instances better be up to date.
Also, use ctx.get_instance, rather than things from the deployment
modification object, because the deployment modification node
instances aren't _exactly_ node instances, they're missing status
entirely.

* RD-6652 Scale: use rollback for rollback

Now that we have a rollback, we can use it instead of
just an uninstall.
This will make it so that rolling back an instance that failed
creating, will only run delete, and not e.g. stop

* Allow `ignore_failure` for rollback

It is passed in from scale, so why not.