Cleanup: drop CloudifySystemWideWorkflowContext
We note that there's a few layers, but at the end, the only difference
between the regular WorkflowContext and a SystemWideWorkflowContext
is that one sends logs with `blueprint_id: ctx.blueprint.id`, and the
other with `blueprint_id: None` (and deployment_id).

Alas, `ctx.blueprint.id` is already None for system-wide executions.
So they are exactly the same, so we can simplify this.