RD-6588 6.4.2 backports (#1222)
* RD-6652 Scale: don't run unnecessary operations (#1219)

* Scale: make sure to refresh instances before doing anything

Very refreshing.
The various `lifecycle` methods examine instance state when scheduling
tasks, so the instances better be up to date.
Also, use ctx.get_instance, rather than things from the deployment
modification object, because the deployment modification node
instances aren't _exactly_ node instances, they're missing status
entirely.

* RD-6652 Scale: use rollback for rollback

Now that we have a rollback, we can use it instead of
just an uninstall.
This will make it so that rolling back an instance that failed
creating, will only run delete, and not e.g. stop

* Allow `ignore_failure` for rollback

It is passed in from scale, so why not.

* RD-6634 Local workflows: cope with adding instances in update (#1210)

When running update, the local storage stuff breaks when adding
new instances. It's mostly unrelated to anything manager, but it
does make testing hard.

- make sure to return NodeInstance wherever a rest-client would
- ...but accept dicts still (no .id), just like the rest-client would
  (dsl-parser's modify_deployment task does return dicts)
- cope with missing 'version'. That is kinda... ugly, but well.

* RD-6285 install: allow filtering node-instances  (#1172)

* `@workflow_test`: allow passing blueprint contents as string

I don't want to always write tests by putting a blueprint in a separate
file, I want to just write it in a string right there! It's going
to make writing tests a lot easier.

Deciding whether the passed-in string is a filename or a blueprint
string is perhaps a bit stupid, but hey, this works.

Some care is taken to make sure all temp files are deleted indeed.

* RD-6285 Install: allow filtering node-instances

Allow installing only a subset of node-instances.

* tests for the install filtering

* flake

* RD-6286 uninstall: allow filtering node-instances (#1173)

Similar to #1172 but for uninstall.

* RD-6287 Declare a `reinstall` workflow (#1174)

* RD-6287 Declare a `reinstall` workflow

Now that `heal` is much more complex, there's no longer a simple
way to "reinstall". We have the `LifecycleProcessor` method right
there, all we need to do is to expose it.

* flake

i'll just NOQA the blueprint, i mean, the operation path is long

* old python compat