RND-297 utils: reverse get_rest_host/port sources
It used to be that the host/port were taken from the context,
and envvars were a fallback.

With this, host/port are taken from the context, but if the envvars
are set, they are _forced_, and override the context values.

This allows some mgmtworkers to be explicitly told, "always connect
to this restservice".

This doesn't affect 7.0 agents, because they (no longer) include
those envvars in their config files.