introduce the `resource:` scheme for locations
not only `file:` is now going to be possible!
`resource:` means that it's an on-manager resource (fileserver
or /opt/manager/resources), to be downloaded via this method,
normally used for fetching things like e.g. types.yaml

Also you'll notice I used `yield` here. Don't worry about it, this
function is not a generator yet, but will be :)