make _get_resource_location into a generator
so the idea is that _get_resource_location is now a generator,
because we can't rely _only_ on the os.path.exists, the urls
need to go through the actual resolver inside of _extract_import_parts
now.

And so, _extract_import_parts has to iterate over it, too.

This is just because we can return multiple 'resource:' entries
(one with _1_3.yaml, one without), and we need to check all of them.

This is a very dirty approach, but I'll clean it up in the next
commits...