RND-430 dsl_parser: download resources from the fileserver (#1266) (#1267)
* move nested functions outside

* inline the other nested functions

also fix the bug in import_url for basic types

* replace recursion with iteration

trivial

* simplify import_key

since it was wrong before, we can surely simplify it

* trivial refactors

* drop pointless location() calls

this is never called with an empty dsl_location, so this is unnecessary

* rewrite the graph sorting using BFS

this is possible because imports are visited in dependency order
anyway, obviously

* refactors - dataclass

* put .get_imports in the dataclass

* build_ordered_imports: only return 2 items

blueprints and mapping contain the same things, we can simplify
by only returning one of them. Also Holder's API allows us to write
these functions a bit shorter.

The tests need to be updated wrt. only returning 2 items

* introduce the `resource:` scheme for locations

not only `file:` is now going to be possible!
`resource:` means that it's an on-manager resource (fileserver
or /opt/manager/resources), to be downloaded via this method,
normally used for fetching things like e.g. types.yaml

Also you'll notice I used `yield` here. Don't worry about it, this
function is not a generator yet, but will be :)

* url_exists: more defensive

this does actually break on things that it shouldnt otherwise,
eg the file: "urls"

* make _get_resource_location into a generator

so the idea is that _get_resource_location is now a generator,
because we can't rely _only_ on the os.path.exists, the urls
need to go through the actual resolver inside of _extract_import_parts
now.

And so, _extract_import_parts has to iterate over it, too.

This is just because we can return multiple 'resource:' entries
(one with _1_3.yaml, one without), and we need to check all of them.

This is a very dirty approach, but I'll clean it up in the next
commits...

* factor out the inner loop

Not a very important refactor, because it moves the thing pretty
much verbatim, but at least now _build_ordered_imports clearly
shows the BFS

* py3.6 compat :(

* add some comments

---------

Co-authored-by: Łukasz Maksymczuk <lukasz@lukasz-maksymczuk.intra.codilime.com>