Sync 7.0.1 with master (#1295)
* RND-364 Implement cloudify_rest_client dump() methods  (#1272)

* Define a constant and get_all/dump_all/dump_blobs functions
* Implement dump() methods for cloudify_rest_client

* RND-364 Don't makedirs for dumping events (#1273)

These are created in snapshot_create system workflow (similarly to the
manner directories are created for other entities)

* RND-364 Snapshot creation: return as many IDs as possible (#1276)

* RND-623 Delete dep-group with deployments: return excgroup id (#1279)

In order to follow the progress of deleting the deployments, it'll
be useful to receive the execution group id

* RND-619 Remove snapshot file creation from rest client (#1280)

Instead .dump() methods return generators which will be used in
snapshot_create workflow.

* RND-623 Add "recursive" for deleting deployments & groups (#1281)

Recursive means that all services contained in the given deployments
(i.e. in the given environments) will be deleted as well.

* Fix how tasks_graphs are dumped (#1283)

* RND-622 uninstall workflow: uninstall service deployments recursively (#1282)

* RND-622 uninstall workflow: uninstall service deployments recursively

With this flag, before uninstalling the target deployment, first
uninstall all deployments that are using it as an environment.

* replace the `while True`

now this will honor cancels

* RD-7086 evaluate nodes: provide an instance id hint (#1284)

* when evaluating nodes, send in the instance id as well

* evaluating functions: use the provided instance id

* Simplify the inline script heuristic (#1285)

The old approach was clever, but still resulted in false positives:
plugin names in blueprints don't actually have to be python
identifiers, so a plugin path like `a-b.c.d.e` ended up being treated
as a script, which is invalid.

I'll just give up, and only leave in the line count approach. Now
a single-line script is disallowed, but that wasn't a very important
angle anyway.

* Implement rest client `restore(entities)` methods (#1286)

* RND-707 Include provider and provider_options in secrets' snapshots (#1287)

* RND-309 Extend rest client's dump() methods (#1288)

Make it possible to dump only specific entities, selected by additional
iterable parameter (e.g. `agent_ids` in case of `AgentsClient` etc.) to
`dump()` functions.

And also add function docstrings.

* RND-673 Include executions.is_system_workflow in snapshots (#1289)

* RND-664 Snapshot restore: log errors and continue (#1290)

* Rest clients snapshot restore fixes (#1291)

* RND-561 scale workflow: add a unique prefix to graph names (#1292)

If this workflow is called multiple times from a custom workflow,
the created graphs need to have unique prefixes, otherwise they won't
run

* RND-652 restclient: add _include to excgroups.get (#1293)

Just declare the parameter

* CYBl-2049 _format_idds: don't break on nodes with 0 instances (#1294)

Nodes can have 0 instances, in which case this would otherwise throw
a KeyError.
Anyway, when there's no instances, there's no IDDs (yet) to speak of.

---------

Co-authored-by: Mateusz Neumann <mateumann@gmail.com>