NE-19591 Mark not-complete executions as cancelled (#1323)
Executions which were still running (i.e. had one of the statuses:
`pending`, `started`, `cancelling`, `force_cancelling` or
`kill_cancelling`) when the snapshot was made, will be marked as
`cancelled` and information about that modification will be save in the
`error` column of the execution.  It means a manual examination may be
required to determine their state.